IoT Servo Motor Source Code/requirements.txt
RPi.GPIO==0.7.1
evdev==1.6.1
matplotlib==3.7.1